+++ Backend/database.sql
@@ -52,6 +52,7 @@ CREATE TABLE identities (
   "sex" varchar(6) CHECK ("sex" IN ('male', 'female')),
   "ethnicity" varchar(12) CHECK ("ethnicity" IN ('caucasian', 'black', 'hispanic', 'latino', 'arab', 'east asian', 'south asian')),
   "email" varchar(48),
+  "email_password" varchar(33),
   "phone" varchar(12),
   "card" varchar(16),
   "status" varchar(8) CHECK ("status" IN ('active', 'inactive', 'frozen')) DEFAULT 'inactive'